Add changing of status and remove unused code
